fix: make sure tracking calls cross the bridge in intended order (#21)
If Matomo is initialized in module scope calls across the bridge are not called in intended order. For example `trackView` was called before  `trackEvent`.

This is more obvious if you set custom dimensions or similar data during initialization that data could get lost or attached to wrong tracking events. Note that custom dimensions are yet to be implemented.